Move node accessors to their new home as vocabulary
