+++ dev/vocabulary/node_algorithms.h
@@ -5,5 +5,6 @@
 
 #include "algorithms/field_predicates_fwd.h"
 #include "algorithms/field_predicates_concepts.h"
-#include "algorithms/predicates.h"
 #include "algorithms/index_filters.h"
+#include "algorithms/predicates.h"
+#include "algorithms/accessors.h"
